+++ index.html
@@ -230,6 +230,10 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 <script src="https://cdn.jsdelivr.net/npm/lodash@4.17.21/lodash.min.js"></script>
-<script defer src="js/main.js"></script>
+<script defer src="js/main.js"  type = "module"></script>
+<script src="/js/utils/data.js" type = "module"></script>
+<script src="/js/utils/create-comment.js" type = "module" ></script>
+<script src="/js/utils/create-object.js" type = "module"></script>
+<script src="/js/utils/generate-objects.js" type = "module"></script>
 </body>
 </html>
