+++ js/utils/data.js
@@ -1,0 +1,23 @@
+const OBJESTS_TO_GEN = 25;
+const MESSAGES = [
+  'Всё отлично!', 'В целом всё неплохо. Но не всё',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра', 'В конце концов это просто непрофессионально.', 'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+const NAMES = ['Ирина', 'Ибрагим', 'Олег', 'Камилла'];
+const LIKES_MIN = 15;
+const LIKES_MAX = 200;
+const COMMENT_MAX = 7;
+const AVATAR_MIN = 1;
+const AVATAR_MAX = 6;
+
+export {
+  OBJESTS_TO_GEN,
+  MESSAGES,
+  NAMES,
+  LIKES_MIN,
+  LIKES_MAX,
+  COMMENT_MAX,
+  AVATAR_MIN,
+  AVATAR_MAX
+};
